Trabajo_practico_2/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.1.3
pillow==11.0.0
propcache==0.2.0
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
